requirements.txt
pyFirmata==1.1.0
pyserial==3.5
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyqt5-tools==5.15.4.3.2